backend/requirements.txt
Django==4.1
djangorestframework>=3.14.0,<3.15
psycopg2>=2.9.9,<2.10
drf-spectacular>=0.26.5,<0.27
djangorestframework_simplejwt>=5.3.1,<5.4
drf-nested-routers>=0.93.4,<0.94